+++ src/contextus.cpp
@@ -73,9 +73,7 @@ struct Contextus : SanguineModule {
 		ENUMS(LIGHT_MODEL, 1 * 3),
 		LIGHT_MORSE,
 		LIGHT_VCA,
-		LIGHT_DRIFT,
 		LIGHT_FLAT,
-		LIGHT_SIGN,
 		LIGHT_AUTO,
 		ENUMS(LIGHT_CHANNEL_MODEL, 16 * 3),
 		LIGHTS_COUNT
@@ -89,6 +87,8 @@ struct Contextus : SanguineModule {
 	renaissance::Quantizer quantizer[PORT_MAX_CHANNELS];
 
 	uint8_t currentScale[PORT_MAX_CHANNELS] = {};
+	uint8_t waveShaperValue = 0;
+	uint8_t driftValue = 0;
 
 	int16_t previousPitch[PORT_MAX_CHANNELS] = {};
 
@@ -109,9 +109,7 @@ struct Contextus : SanguineModule {
 	bool bTriggerFlag[PORT_MAX_CHANNELS] = {};
 
 	bool bAutoTrigger = false;
-	bool bDritfEnabled = false;
 	bool bFlattenEnabled = false;
-	bool bSignatureEnabled = false;
 	bool bVCAEnabled = false;
 
 	bool bLowCpu = false;
@@ -170,9 +168,9 @@ struct Contextus : SanguineModule {
 		// Unused: kept for compatibility.
 		configButton(PARAM_META, "");
 		configButton(PARAM_VCA, "Toggle AD VCA");
-		configButton(PARAM_DRIFT, "Toggle oscillator drift");
+		configSwitch(PARAM_DRIFT, 0.f, 4.f, 0.f, "Oscillator drift", nodiIntensityTooltipStrings);
 		configButton(PARAM_FLAT, "Toggle lower and higher frequency detuning");
-		configButton(PARAM_SIGN, "Toggle output imperfections");
+		configSwitch(PARAM_SIGN, 0.f, 4.f, 0.f, "Signature wave shaper", nodiIntensityTooltipStrings);
 		configButton(PARAM_AUTO, "Toggle auto trigger");
 
 		for (int channel = 0; channel < PORT_MAX_CHANNELS; ++channel) {
@@ -206,10 +204,10 @@ struct Contextus : SanguineModule {
 		channelCount = std::max(std::max(inputs[INPUT_PITCH].getChannels(), inputs[INPUT_TRIGGER].getChannels()), 1);
 
 		bVCAEnabled = params[PARAM_VCA].getValue();
-		bDritfEnabled = params[PARAM_DRIFT].getValue();
 		bFlattenEnabled = params[PARAM_FLAT].getValue();
-		bSignatureEnabled = params[PARAM_SIGN].getValue();
 		bAutoTrigger = params[PARAM_AUTO].getValue();
+		waveShaperValue = params[PARAM_SIGN].getValue();
+		driftValue = params[PARAM_DRIFT].getValue();
 
 		outputs[OUTPUT_OUT].setChannels(channelCount);
 
@@ -259,9 +257,9 @@ struct Contextus : SanguineModule {
 			// Handle switches
 			settings[channel].meta_modulation = 1;
 			settings[channel].ad_vca = bVCAEnabled;
-			settings[channel].vco_drift = bDritfEnabled;
+			settings[channel].vco_drift = driftValue;
 			settings[channel].vco_flatten = bFlattenEnabled;
-			settings[channel].signature = bSignatureEnabled;
+			settings[channel].signature = waveShaperValue;
 			settings[channel].auto_trig = bAutoTrigger;
 
 			// Quantizer
@@ -472,27 +470,27 @@ struct Contextus : SanguineModule {
 			}
 			case renaissance::SETTING_AD_ATTACK: {
 				value = settings[displayChannel].ad_attack;
-				displayText = nodiNumberStrings[value];
+				displayText = nodiNumberStrings15[value];
 				break;
 			}
 			case renaissance::SETTING_AD_DECAY: {
 				value = settings[displayChannel].ad_decay;
-				displayText = nodiNumberStrings[value];
+				displayText = nodiNumberStrings15[value];
 				break;
 			}
 			case renaissance::SETTING_AD_FM: {
 				value = settings[displayChannel].ad_fm;
-				displayText = nodiNumberStrings[value];
+				displayText = nodiNumberStrings15[value];
 				break;
 			}
 			case renaissance::SETTING_AD_TIMBRE: {
 				value = settings[displayChannel].ad_timbre;
-				displayText = nodiNumberStrings[value];
+				displayText = nodiNumberStrings15[value];
 				break;
 			}
 			case renaissance::SETTING_AD_COLOR: {
 				value = settings[displayChannel].ad_color;
-				displayText = nodiNumberStrings[value];
+				displayText = nodiNumberStrings15[value];
 				break;
 			}
 			case renaissance::SETTING_AD_VCA: {
@@ -524,11 +522,13 @@ struct Contextus : SanguineModule {
 				break;
 			}
 			case renaissance::SETTING_VCO_DRIFT: {
-				displayText = nodiDriftLabel;
+				value = settings[displayChannel].vco_drift;
+				displayText = nodiIntensityDisplayStrings[value];
 				break;
 			}
 			case renaissance::SETTING_SIGNATURE: {
-				displayText = nodiSignLabel;
+				value = settings[displayChannel].signature;
+				displayText = nodiIntensityDisplayStrings[value];
 				break;
 			}
 			default: {
@@ -558,9 +558,7 @@ struct Contextus : SanguineModule {
 	inline void pollSwitches(const float sampleTime) {
 		// Handle switch lights
 		lights[LIGHT_VCA].setBrightnessSmooth(bVCAEnabled ? 0.75f : 0.f, sampleTime);
-		lights[LIGHT_DRIFT].setBrightnessSmooth(bDritfEnabled ? 0.75f : 0.f, sampleTime);
 		lights[LIGHT_FLAT].setBrightnessSmooth(bFlattenEnabled ? 0.75f : 0.f, sampleTime);
-		lights[LIGHT_SIGN].setBrightnessSmooth(bSignatureEnabled ? 0.75f : 0.f, sampleTime);
 		lights[LIGHT_AUTO].setBrightnessSmooth(bAutoTrigger ? 0.75f : 0.f, sampleTime);
 	}
 
@@ -649,15 +647,14 @@ struct ContextusWidget : SanguineModuleWidget {
 
 		addParam(createParamCentered<Sanguine1PSGreen>(millimetersToPixelsVec(119.474, 36.666), module, Contextus::PARAM_ATTACK));
 
-		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(22.768, 54.243), module, Contextus::PARAM_AD_TIMBRE));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(119.4, 54.231),
+		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(35.151, 45.206), module, Contextus::PARAM_AD_TIMBRE));
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(107.074, 45.206),
 			module, Contextus::PARAM_VCA, Contextus::LIGHT_VCA));
 
 		addParam(createParamCentered<Sanguine1PSPurple>(millimetersToPixelsVec(10.076, 67.247), module, Contextus::PARAM_MODULATION));
 		addParam(createParamCentered<Sanguine1PSRed>(millimetersToPixelsVec(36.032, 67.247), module, Contextus::PARAM_ROOT));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<YellowLight>>>(millimetersToPixelsVec(48.572, 80.197),
-			module, Contextus::PARAM_DRIFT, Contextus::LIGHT_DRIFT));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<OrangeLight>>>(millimetersToPixelsVec(93.673, 80.197),
+		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(62.607, 105.967), module, Contextus::PARAM_DRIFT));
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<OrangeLight>>>(millimetersToPixelsVec(71.12, 96.625),
 			module, Contextus::PARAM_FLAT, Contextus::LIGHT_FLAT));
 		addParam(createParamCentered<Sanguine1PSRed>(millimetersToPixelsVec(106.234, 67.247), module, Contextus::PARAM_SCALE));
 		addParam(createParamCentered<Sanguine1PSGreen>(millimetersToPixelsVec(132.166, 67.247), module, Contextus::PARAM_DECAY));
@@ -669,8 +666,7 @@ struct ContextusWidget : SanguineModuleWidget {
 		addParam(createParamCentered<Sanguine1PSBlue>(millimetersToPixelsVec(22.768, 97.889), module, Contextus::PARAM_COLOR));
 
 		addParam(createParamCentered<Sanguine1PSRed>(millimetersToPixelsVec(51.457, 93.965), module, Contextus::PARAM_PITCH_OCTAVE));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<PurpleLight>>>(millimetersToPixelsVec(71.12, 93.962),
-			module, Contextus::PARAM_SIGN, Contextus::LIGHT_SIGN));
+		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(79.633, 105.967), module, Contextus::PARAM_SIGN));
 		addParam(createParamCentered<Sanguine1PSRed>(millimetersToPixelsVec(90.806, 93.965), module, Contextus::PARAM_PITCH_RANGE));
 
 		addParam(createParamCentered<Sanguine1PSOrange>(millimetersToPixelsVec(119.474, 97.889), module, Contextus::PARAM_FM));
@@ -681,8 +677,8 @@ struct ContextusWidget : SanguineModuleWidget {
 		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(35.151, 117.788), module, Contextus::PARAM_TRIGGER_DELAY));
 		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(46.798, 117.788),
 			module, Contextus::PARAM_AUTO, Contextus::LIGHT_AUTO));
-		addParam(createParamCentered<Sanguine1PSYellow>(millimetersToPixelsVec(62.4, 113.511), module, Contextus::PARAM_BITS));
-		addParam(createParamCentered<Sanguine1PSYellow>(millimetersToPixelsVec(79.841, 113.511), module, Contextus::PARAM_RATE));
+		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(62.607, 118.103), module, Contextus::PARAM_BITS));
+		addParam(createParamCentered<Trimpot>(millimetersToPixelsVec(79.633, 118.103), module, Contextus::PARAM_RATE));
 		addOutput(createOutputCentered<BananutRedPoly>(millimetersToPixelsVec(133.968, 117.788), module, Contextus::OUTPUT_OUT));
 
 		SanguineBloodLogoLight* bloodLogo = new SanguineBloodLogoLight(module, 98.491, 112.723);